Rename utils.ts file name
