+++ app/home/component/MovieCard/MovieCard.utils.ts
