+++ app/home/page.tsx
@@ -1,6 +1,6 @@
 import { fetchPopularMovies } from "../lib/http/movie/movie.service";
 
-import { convertMovieResponseToCard } from "./component/MovieCard/utils";
+import { convertMovieResponseToCard } from "./component/MovieCard/MovieCard.utils";
 import { MovieCardListProps, MovieCardList } from "./component/MovieCardList/MovieCardList";
 
 export default async function Home() {
